+++ NEWS.md
@@ -1,5 +1,12 @@
 \title{ISAnalytics News}
 
+# ISAnalytics 1.1.4 (2020-11-16)
+
+## UPGRADES
+
+* Greatly improved reports for collision removal function
+* General improvements for all widget reports
+
 # ISAnalytics 1.1.3 (2020-11-10)
 
 ## FIXES